refactor: rename cmake_settings to skbuild_settings
